requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
openapi-codec==1.3.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
